requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
pymysql
pymongo
google.generativeai